pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "havona-sdk"
version = "0.1.0"
description = "Python client for the Havona trade finance API"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
keywords = ["trade-finance", "blockchain", "etr", "havona", "api-client"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/havona-labs/havona-sdk"
Documentation = "https://github.com/havona-labs/havona-sdk#readme"
"Issue Tracker" = "https://github.com/havona-labs/havona-sdk/issues"

[tool.hatch.build.targets.wheel]
packages = ["havona_sdk"]
